[ch]-move validation folder to client side
